+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## Added
 
+- Added `KVMWorkerHostVolumes` in worker template.
+
+## Added
+
 - Added `enableServer` config in kubelet config file
 - Propagate proxy environments to the node templates.